Done with spinner  in sidebar (#39)
